Fix: TaskRepositoryの依存性を解消
