+++ lib/main.dart
@@ -2,7 +2,9 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_dynamic_links/firebase_dynamic_links.dart';
 import 'package:flutter/material.dart';
 import 'package:package_info/package_info.dart';
+import 'package:projecthit/implements/task_repository_impl.dart';
 import 'package:projecthit/model/theme_model.dart';
+import 'package:projecthit/repository/task_repository.dart';
 import 'package:projecthit/screens/accept_invitation/accept_invitation_page.dart';
 import 'package:projecthit/screens/my_app/my_app_page.dart';
 import 'package:provider/provider.dart';
@@ -35,6 +37,9 @@ void main() async {
       providers: [
         Provider<PackageInfo>.value(value: packageInfo),
         ChangeNotifierProvider<ThemeModel>(create: (_) => themeModel),
+        Provider<TaskRepository>(
+          create: (_) => TaskRepositoryImpl(),
+        ),
       ],
       child: MyApp(navigatorKey: navigatorKey),
     ),
